+++ src/types/index.ts
@@ -1,3 +1,4 @@
+export * from "./auth";
 export * from "./base";
 export * from "./app";
 export * from "./page";